--- app/src/main/java/com/mashup/molink/interest/Data.kt
@@ -1,24 +1,0 @@
-package com.mashup.molink.interest
-
-import com.google.gson.annotations.Expose
-import com.google.gson.annotations.SerializedName
-
-
-
-
-data class Data(
-
-    @SerializedName("id")
-    @Expose
-    val id: Int? = null,
-
-    @SerializedName("imgUrl")
-    @Expose
-    val imgUrl: String? = null,
-
-    @SerializedName("name")
-    @Expose
-    val name: String? = null,
-
-    var check: Boolean = false
-)
