--- app/src/main/java/com/mashup/molink/interest/Interest.kt
@@ -1,6 +1,0 @@
-package com.mashup.molink.interest
-
-import android.graphics.drawable.Drawable
-
-data class Interest(val name: String, var check: Boolean, val src: Drawable?)
-
