+++ app/src/main/java/com/mashup/molink/ui/interest/Interest.kt
@@ -1,0 +1,6 @@
+package com.mashup.molink.ui.interest
+
+import android.graphics.drawable.Drawable
+
+data class Interest(val name: String, var check: Boolean, val src: Drawable?)
+
