+++ app/src/main/res/layout/activity_login.xml
@@ -6,7 +6,7 @@
         android:background="@color/white_gray"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".login.LoginActivity">
+        tools:context=".ui.login.LoginActivity">
 
     <LinearLayout android:layout_width="match_parent" android:layout_height="wrap_content"
                   android:background="@drawable/background_circle"
