+++ app/src/main/res/layout/activity_register.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".register.RegisterActivity"
+        tools:context=".ui.register.RegisterActivity"
         android:background="@color/white_gray">
 
     <androidx.appcompat.widget.AppCompatImageView android:layout_width="130dp" android:layout_height="130dp"
